+++ src/app/app.component.ts
@@ -1,17 +1,23 @@
-import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, ViewChild, inject } from '@angular/core';
+import { ConstantsService, constantsService } from './core/services/constants.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.css'],
+  providers: [
+    { provide: ConstantsService, useValue: constantsService },
+  ]
 })
 export class AppComponent implements AfterViewInit {
   title = 'shop';
+
+  private constantsService = inject(ConstantsService);
 
   @ViewChild("appTitle")
   appTitle!: ElementRef<HTMLHeadingElement>;
 
   ngAfterViewInit(): void {
-    this.appTitle.nativeElement.textContent = "The best shop ever";
+    this.appTitle.nativeElement.textContent = "The best shop ever! " + this.constantsService.ApiUrl;
   }
 }