+++ src/app/cart/models/cart-item.model.ts
@@ -1,14 +1,6 @@
 import { ProductModel } from "src/app/products/models/product.model";
 
 export class CartItemModel {
-    constructor(public product: ProductModel, public count: number = 1) {
-    }
-
-    incrementCount(): void {
-        this.count++;
-    }
-
-    decrementCount(): void {
-        this.count--;
+    constructor(public product: ProductModel, public quantity: number = 1) {
     }
 }