+++ src/app/core/models/config.model.ts
@@ -1,0 +1,6 @@
+export class ConfigModel {
+    constructor(
+        public id: number,
+        public login: string,
+        public email: string) {}
+}