+++ src/app/core/services/generator.factory.ts
@@ -1,0 +1,6 @@
+import { GeneratorService } from "./generator.service";
+
+export const GeneratorFactory = function (generator: GeneratorService): (n: number) => string
+{
+    return (length) => generator.generate(length);
+}