+++ src/app/products/components/product-list/product-list.component.ts
@@ -20,6 +20,6 @@ export class ProductListComponent implements OnInit {
   }
 
   onProductToCartAdded(product: ProductModel) {
-    this.cartService.addToCart(product);
+    this.cartService.addProduct(product);
   }
 }