+++ src/app/products/products.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ProductComponent } from './components/product/product.component';
 import { ProductListComponent } from './components/product-list/product-list.component';
+import { SharedModule } from '../shared/shared.module';
 
 
 
@@ -11,10 +12,10 @@ import { ProductListComponent } from './components/product-list/product-list.com
     ProductListComponent,
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    SharedModule,
   ],
   exports: [
-    // ProductComponent,
     ProductListComponent,
   ]
 })