+++ src/app/shared/directives/click-to-highlight.directive.spec.ts
@@ -1,0 +1,8 @@
+import { ClickToHighlightDirective } from './click-to-highlight.directive';
+
+describe('ClickToHighlightDirective', () => {
+  it('should create an instance', () => {
+    const directive = new ClickToHighlightDirective();
+    expect(directive).toBeTruthy();
+  });
+});